pyproject.toml
[project]
name = "analytics-interview"
version = "0.1.0"
dependencies = [
    "dbt-postgres~=1.8.0",
    "duckdb==1.4.1",
    "pyyaml==6.0.3"
]
